Refactor Spectrogram to use MVVM pattern
