+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/MeasurementModule.kt
@@ -3,6 +3,7 @@ package org.noiseplanet.noisecapture.ui.features.measurement
 import org.koin.compose.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 import org.noiseplanet.noisecapture.ui.features.measurement.indicators.AcousticIndicatorsViewModel
+import org.noiseplanet.noisecapture.ui.features.measurement.spectrogram.SpectrogramPlotViewModel
 import org.noiseplanet.noisecapture.ui.features.measurement.spectrum.SpectrumPlotViewModel
 
 
@@ -14,5 +15,9 @@ val measurementModule = module {
 
     viewModel {
         SpectrumPlotViewModel(measurementsService = get())
+    }
+
+    viewModel {
+        SpectrogramPlotViewModel(measurementsService = get())
     }
 }
